use GKeyFile for loading/storing the API key
This patch uses the GKeyFile API to store the provided API key through the
command line. If no API key is set, then it is tried to be loaded, assuming
that the key was stored in the past.

The configuration file path is XDG user config one (~/.config).

This behavior is more easy to use since no user interventions is required to
create the configuration file.

Signed-off-by: Víctor Manuel Jáquez Leal <vjaquez@igalia.com>