ui: make dbus call in FeedServer async
FeedServer must not block the GUI. So this change implements and asynchronous call
using GLib main loop.

FeedServer.search() method has a new optional parameter with the optional callback
function that will be invoked on D-Bus method response.

It also adds a new test_async() to show the callback use.